+++ src/main/java/com/musinsa/orders/domain/refund/RefundPolicy.java
@@ -1,5 +1,7 @@
-package com.musinsa.orders.domain;
+package com.musinsa.orders.domain.refund;
 
+import com.musinsa.orders.domain.Money;
+import com.musinsa.orders.domain.Order;
 import java.util.List;
 
 public class RefundPolicy {
