+++ src/test/java/com/musinsa/orders/domain/RefundPolicyTest.java
@@ -4,6 +4,7 @@ import static com.musinsa.orders.Fixtures.createOrder;
 import static com.musinsa.orders.Fixtures.createOrderLineItem;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.musinsa.orders.domain.refund.RefundPolicy;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
